Implement lazy execution logic with lazy resolvers
To mitigate the N+1 queries problem I introduced the concept of lazy
resolvers. Lazy resolvers are classes, instances of which are
responsible for resolution of the GraphQL fields affected by N+1
queries. This is how it works:

The `Spree::GraphQL::LazyResolver::TypeHelper` module is included within
GraphQL classes that require lazy execution logic. Such module defines a
`lazy_resolver` protected method, to which the GraphQL type class
delegates the resolution of the involved fields. The `lazy_resolver`
method is responsible for finding the lazy resolver class associated to
the selected lazy resolver adapter (more on the lazy resolver adapter
below) and to the involved GraphQL type class. After the lazy resolver
class is found, it gets initialized with the GraphQL type instance
object and the GraphQL context, and the resolution of the GraphQL field
is delegated to the created instance.

To let developers pick the lazy execution logic most fitting for them,
or even implement their ones, LazyResolver uses the Adapter pattern. The
default adapter (and the only one implemented at the moment) is for the
`batch-loader` gem. I chose `batch-loader` over the many alternatives
(`graphql-batch`, `graphql` vanilla lazy resolvers, ...) for the
following reasons:
- adequately popular
- maintained
- it implements the lazy object pattern, which is more idiomatic in the
  Ruby language context than the promise pattern, implemented by some of
  the other popular alternatives (`graphql-batch`, `dataloader`)
- I found it easy to use