Change decorator file structure
This is not an urgent change, but for the sake of establishing this
pattern before more decorators are added, I think it should be done now.

I prefer this way of organizing decorators because it makes it more
obvious to the developer which classes are intended to be decorators and
which are completely new models or controllers.